Add "Auto Gen XML" check box
"Auto Gen" check box when ticked does an XML Generation when any of the CFG Group fields change. This saves having to Click that button after each change to see the effect.

Reordered the Checks so "Auto Gen" is 1st and "No !!!" is last to use as Range of FEQs... not the most solid code.